Corrected SSR Stats Bug, plus a few formatting details
